Many changes

+++ controller/src/main/java/by/xenon28082/shop/controller/commands/commandsImpl/DeleteFromBasketCommand.java
@@ -1,8 +1,10 @@
 package by.xenon28082.shop.controller.commands.commandsImpl;
 
 import by.xenon28082.shop.controller.commands.Command;
+import by.xenon28082.shop.controller.exception.ControllerException;
 import by.xenon28082.shop.controller.validators.Validator;
 import by.xenon28082.shop.controller.validators.ValidatorImpl;
+import by.xenon28082.shop.entity.Product;
 import by.xenon28082.shop.service.OrderService;
 import by.xenon28082.shop.service.ProductService;
 import by.xenon28082.shop.service.ServiceFactory;
@@ -34,7 +36,7 @@ public class DeleteFromBasketCommand implements Command {
     private static final String PRODUCT_AMOUNT = "productAmount";
 
     @Override
-    public void execute(HttpServletRequest req, HttpServletResponse res) throws SQLException, ServletException, IOException, ServiceException {
+    public void execute(HttpServletRequest req, HttpServletResponse res) throws ControllerException {
         LOGGER.info("Got to DeleteFromBasketCommand");
 
         final long id = (Long) req.getSession().getAttribute(ID);
@@ -42,22 +44,34 @@ public class DeleteFromBasketCommand implements Command {
         final long productId = Long.parseLong(req.getParameter(PRODUCT_ID));
         final int amountToDelete = Integer.parseInt(req.getParameter(PRODUCT_AMOUNT));
 
-        boolean hasNotDeleted = orderService.findReservation(productId, id).getAmount() != 0;//productService.findProductById(productId).getStock() != 0;
+        boolean hasNotDeleted = false;
+        try {
+            hasNotDeleted = orderService.findReservation(productId, id).getAmount() != 0;
 
-        if (hasNotDeleted && validator.validateIsNotPositive(
-                Arrays.asList(String.valueOf(amountToDelete))
-        )) {
-            LOGGER.info("Value is not positive");
-            res.sendRedirect("FrontController?COMMAND=SHOW_BASKET&message=failed");
-        } else {
-            boolean isDeleted = orderService.deleteReservation(id, orderId, amountToDelete, hasNotDeleted);
-            productService.updateProduct(productId, -amountToDelete);
-            if (isDeleted) {
-                LOGGER.info("Delete order: id - " + id + " orderId - " + orderId + " toDelete - " + amountToDelete + " (SUCCESS)");
+
+            if (hasNotDeleted && validator.validateIsNotPositive(
+                    Arrays.asList(String.valueOf(amountToDelete))
+            )) {
+                LOGGER.info("Value is not positive");
+                res.sendRedirect("FrontController?COMMAND=SHOW_BASKET&message=failed");
             } else {
-                LOGGER.info("Delete order: id - " + id + " orderId - " + orderId + " toDelete - " + amountToDelete + " (FAILED)");
+                Product product = productService.findProductById(productId);
+                if (product.getStock() < amountToDelete) {
+                    LOGGER.info("Can't delete more then have");
+                    res.sendRedirect("FrontController?COMMAND=SHOW_BASKET&message=more");
+                } else {
+                    boolean isDeleted = orderService.deleteReservation(id, orderId, amountToDelete, hasNotDeleted);
+                    productService.updateProduct(productId, -amountToDelete);
+                    if (isDeleted) {
+                        LOGGER.info("Delete order: id - " + id + " orderId - " + orderId + " toDelete - " + amountToDelete + " (SUCCESS)");
+                    } else {
+                        LOGGER.info("Delete order: id - " + id + " orderId - " + orderId + " toDelete - " + amountToDelete + " (FAILED)");
+                    }
+                    res.sendRedirect("FrontController?COMMAND=SHOW_BASKET");
+                }
             }
-            res.sendRedirect("FrontController?COMMAND=SHOW_BASKET");
+        } catch (ServiceException | IOException e) {
+            throw new ControllerException(e);
         }
     }
